refactor(debug): Use `debug-js`
