+++ deps/debug.ts
@@ -1,0 +1,1 @@
+export * from "https://raw.githubusercontent.com/takker99/debug-js/0.1.0/mod.ts";
