+++ mod.ts
@@ -2,14 +2,13 @@
 /// <reference lib="esnext" />
 /// <reference lib="dom" />
 
-import { createDebug } from "./debug.ts";
+import { createDebug } from "./deps/debug.ts";
 import { downloadLinks } from "./remote.ts";
 import { fetchProjectStatus, ProjectStatus } from "./status.ts";
 import { open, Source, SourceStatus, write } from "./db.ts";
 import { emitChange } from "./subscribe.ts";
 export { subscribe } from "./subscribe.ts";
 export type { LinkEvent, Listener } from "./subscribe.ts";
-export { setDebugMode } from "./debug.ts";
 export type { Source };
 export * from "./link.ts";
